+++ helpertypes.c
@@ -1,6 +1,6 @@
 #include "helpertypes.h"
 coords* __attribute__((pure)) ringbuffer_getoffset (const ringbuffer* const input,const int offset)
 {
-    if (offset <= input->curidx) {return input->data[input->curidx - offset];}
-    else {return input->data[input->curidx - offset + input->count];}
+    if (offset <=(int) input->curidx) {return input->data[(int)input->curidx - offset];}
+    else {return input->data[(int)input->curidx - offset + (int)input->count];}
 }
